refactor: extract to own component
